mkwdashboard/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dash==2.18.2
numpy==2.2.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
pycollada==0.8
